Wire up background script router
