+++ apps/extension/src/background/index.ts
@@ -1,21 +1,20 @@
-import { init, AEAD, Bip44, Mnemonic, PhraseSize } from "@anoma/crypto";
+import { init } from "@anoma/crypto";
+import { ExtensionRouter } from "../extension";
+import { ContentScriptEnv } from "../utils";
+import { Ports } from "../router/types";
+import { AddChainMsg } from "../router";
 
 const initWasm = async () => {
   await init();
-  console.log(PhraseSize.Twelve, PhraseSize.TwentyFour);
-  const password = "My_Very_$Ecure_PAssW0rd!";
-  const encrypted = AEAD.encrypt("Encrypt my secret message", password);
-  const decrypted = AEAD.decrypt(encrypted, password);
-  console.log({ encrypted, decrypted });
-
-  const mnemonic = Mnemonic.new(PhraseSize.TwentyFour);
-  console.log({ mnemonic });
-
-  const seed = mnemonic.to_bytes();
-  const path = "m/44'/0'/0'/0";
-  const bip44 = Bip44.new(seed);
-  const account = bip44.derive(path);
-  console.log({ bip44, account });
+  console.info("wasm loaded");
 };
 
 initWasm();
+
+const router = new ExtensionRouter(ContentScriptEnv.produceEnv);
+router.registerMessage(AddChainMsg);
+// TODO: This is just an example to test connectivity
+router.addHandler("add-chain", (env, msg) =>
+  console.log("add-chain", { env, msg })
+);
+router.listen(Ports.Background);
