+++ apps/extension/src/router/Router.ts
@@ -52,6 +52,10 @@ export abstract class Router {
     if (!route) {
       throw new Error("Null router");
     }
+    console.log("ExtensionRouter.handleMessage", {
+      route,
+      handlers: this.registeredHandler,
+    });
     const handler = this.registeredHandler.get(route);
     if (!handler) {
       throw new Error("Can't get handler");
