+++ apps/extension/src/utils/index.ts
@@ -1,3 +1,6 @@
+import browser from "webextension-polyfill";
+import { Env, MessageSender } from "../router/types";
+
 declare global {
   var anomaExtensionRouterId: number;
 }
@@ -8,3 +11,19 @@ export const getAnomaRouterId = (): number => {
   }
   return window.anomaExtensionRouterId;
 };
+
+// Determine if content-scripts can be executed in this environment
+export class ContentScriptEnv {
+  static readonly produceEnv = (sender: MessageSender): Env => {
+    const isInternalMsg = sender.id === browser.runtime.id;
+
+    return {
+      isInternalMsg,
+      requestInteraction: () => {
+        throw new Error(
+          "ContentScriptEnv doesn't support `requestInteraction`"
+        );
+      },
+    };
+  };
+}
